16
Chapter 16

Adding Metadata
Metadata is crucial for SEO and shareability. In this chapter, we'll discuss how you can add metadata to your Next.js application.

In this chapter...

Here are the topics we’ll cover

What metadata is.

Types of metadata.

How to add an Open Graph image using metadata.

How to add a favicon using metadata.